refactor(characters): extract Characters.Progression
Sixth slice of #134, and the largest: pending-choice computation, slot
tracking, and option filtering. count_progression_decisions stays
private to Progression — the resolution machinery still in the facade
only needs it through next_progression_decision, which is already part
of the public API. Progression reaches Effects and Leveling directly
rather than through the facade so the dependency between concerns is
explicit.
